giftcard screen issue resolved
