+++ package.json
@@ -42,6 +42,7 @@
     "react-native-snap-carousel": "^3.9.1",
     "react-native-svg": "^12.1.0",
     "react-native-swiper": "^1.6.0",
+    "react-native-swiper-hook": "^1.0.11",
     "react-redux": "^7.2.1",
     "redux": "^4.0.5",
     "redux-persist": "^6.0.0",
