+++ src/components/GiftCard/index.js
@@ -2,3 +2,4 @@ export * from './SubCategory';
 export * from './GiftCardBox';
 export * from './SubAmount';
 export * from './SubUpload';
+export * from './SubGiftCard';
