+++ src/constants/data.js
@@ -2,28 +2,28 @@ import images from 'constants/images';
 
 export default {
   giftCards: [
-    {title: 'Amazon', image: images.amazon},
-    {title: 'American exp', image: images.americanexpress},
-    {title: 'Applestore', image: images.bestbuy},
-    {title: 'colorpop', image: images.ebay},
-    {title: 'ebay', image: images.footlocker},
-    {title: 'itunes', image: images.footlocker},
-    {title: 'JCPenny', image: images.gamestop},
-    {title: 'macy', image: images.gamestop},
-    {title: 'nike', image: images.nike},
-    {title: 'offgamers', image: images.nike},
-    {title: 'Playstation', image: images.nike},
-    {title: 'steam', image: images.nike},
-    {title: 'Bestbuy', image: images.nike},
-    {title: 'sakes', image: images.nike},
-    {title: 'google play', image: images.googleplay},
-    // {title: 'Nordstorm', image: images.nordstrom},
-    {title: 'razer', image: images.razergold},
-    {title: 'sephora', image: images.sephora},
-    {title: 'Target', image: images.target},
-    {title: 'Vanilla', image: images.vanila},
-    {title: 'visa', image: images.visa},
-    // {title: 'walmart', image: images.walmart},
+    {title: 'Amazon', displayName: 'Amazon'},
+    {title: 'American exp', displayName: 'American Express'},
+    {title: 'Applestore', displayName: 'Apple Store'},
+    {title: 'colorpop', displayName: 'colorpop'},
+    {title: 'ebay', displayName: 'ebay'},
+    {title: 'itunes', displayName: 'itunes'},
+    {title: 'JCPenny', displayName: 'JCPenny'},
+    {title: 'macy', displayName: 'macy'},
+    {title: 'nike', displayName: 'nike'},
+    {title: 'offgamers', displayName: 'offgamers'},
+    {title: 'Playstation', displayName: 'Playstation'},
+    {title: 'steam', displayName: 'steam'},
+    {title: 'Bestbuy', displayName: 'Bestbuy'},
+    {title: 'sakes', displayName: 'sakes'},
+    {title: 'google play', displayName: 'Google Play'},
+    // {title: 'Nordstorm', displayName:  'Nordstorm',},
+    {title: 'razer', displayName: 'razer'},
+    {title: 'sephora', displayName: 'sephora'},
+    {title: 'Target', displayName: 'Target'},
+    {title: 'Vanilla', displayName: 'Vanilla'},
+    {title: 'visa', displayName: 'visa'},
+    // {title: 'walmart', displayName:  'walmart',},
     // xbox
   ],
   cardSub: {
