+++ src/screens/GiftCard/GiftCard.js
@@ -1,4 +1,4 @@
-import React, {useRef} from 'react';
+import React, {useMemo, useRef, useState} from 'react';
 import {
   ScrollView,
   Image,
@@ -12,13 +12,23 @@ import Swiper from 'react-native-swiper';
 import Carousel from 'react-native-snap-carousel';
 
 import {Box, Text, Divider, Button, HeaderInfo, Icon} from 'components';
-import {SubCategory, SubUpload, SubAmount} from 'components/GiftCard';
+import {
+  SubGiftCard,
+  SubCategory,
+  SubUpload,
+  SubAmount,
+} from 'components/GiftCard';
 import data from 'constants/data';
+import {capitalizeFirstLetter} from 'shared/utils';
 
-import {RenderItem} from './RenderItem';
+import {RenderItem} from '../../components/GiftCard/RenderItem';
 
 export const GiftCard = () => {
   // let [sliderIndex, setSliderIndex] = useState(0);
+  const [index, setIndex] = useState(0);
+  const [giftCard, setSelected] = useState(null);
+
+  const selectedGiftCard = useMemo(() => data.giftCards[index], [index]);
 
   let navigation = useNavigation();
   // let toSubCategory = () => navigation.navigate('SubCategory');
@@ -29,7 +39,6 @@ export const GiftCard = () => {
   const goBack = () => {
     swiperRef.current?.scrollBy(-1);
   };
-  const CAROUSEL_WIDTH = Dimensions.get('screen').width - 2 * 20;
   return (
     <Box flex={1}>
       <Divider marginBottom="l" />
@@ -37,22 +46,6 @@ export const GiftCard = () => {
         {/* Header */}
         {/* Content */}
 
-        <Box
-          height={250}
-          alignItems="center"
-          marginVertical={{bigScreen: 'xxl', phone: 'xl'}}>
-          <Carousel
-            // ref={(c) => { this._carousel = c; }}
-            data={data.giftCards}
-            layout="stack"
-            loop
-            renderItem={RenderItem}
-            sliderWidth={CAROUSEL_WIDTH}
-            itemWidth={CAROUSEL_WIDTH}
-            containerCustomStyle={styles.slider}
-            contentContainerCustomStyle={styles.sliderContentContainer}
-          />
-        </Box>
         <Swiper
           ref={swiperRef}
           style={giftcardStyles.wrapper}
@@ -62,27 +55,30 @@ export const GiftCard = () => {
           // style={{height: 250}}
           //
         >
-          <Box>
-            <HeaderInfo text="SWIPE THROUGH TO SELECT CATEGORY" />
-
-            <Text
-              color="primary"
-              fontWeight="600"
-              textAlign="center"
-              fontSize={12}
-              marginBottom="m">
-              CLICK HERE TO BEGIN
-            </Text>
-
-            <Button
-              variant="giftcard"
-              text="Ebay"
-              onPress={() => goToNextSlide()}
-            />
-          </Box>
-          <SubCategory next={goToNextSlide} prev={goBack} />
-          <SubAmount next={goToNextSlide} prev={goBack} />
-          <SubUpload next={goToNextSlide} prev={goBack} />
+          <SubGiftCard
+            next={goToNextSlide}
+            prev={goBack}
+            onSnapToItem={(slideIndex) => {
+              setIndex(slideIndex);
+              setSelected(data.giftCards[slideIndex]);
+            }}
+            selectedGiftCard={selectedGiftCard}
+          />
+          <SubCategory
+            data={selectedGiftCard}
+            next={goToNextSlide}
+            prev={goBack}
+          />
+          <SubAmount
+            data={selectedGiftCard}
+            next={goToNextSlide}
+            prev={goBack}
+          />
+          <SubUpload
+            data={selectedGiftCard}
+            next={goToNextSlide}
+            prev={goBack}
+          />
         </Swiper>
       </ScrollView>
       <Box paddingHorizontal="l" marginBottom="l">
@@ -153,65 +149,3 @@ export const colors = {
   background1: '#B721FF',
   background2: '#21D4FD',
 };
-
-const styles = StyleSheet.create({
-  safeArea: {
-    flex: 1,
-    backgroundColor: colors.black,
-  },
-  container: {
-    flex: 1,
-    backgroundColor: colors.background1,
-  },
-  gradient: {
-    ...StyleSheet.absoluteFillObject,
-  },
-  scrollview: {
-    flex: 1,
-  },
-  exampleContainer: {
-    paddingVertical: 30,
-  },
-  exampleContainerDark: {
-    backgroundColor: colors.black,
-  },
-  exampleContainerLight: {
-    backgroundColor: 'white',
-  },
-  title: {
-    paddingHorizontal: 30,
-    backgroundColor: 'transparent',
-    color: 'rgba(255, 255, 255, 0.9)',
-    fontSize: 20,
-    fontWeight: 'bold',
-    textAlign: 'center',
-  },
-  titleDark: {
-    color: colors.black,
-  },
-  subtitle: {
-    marginTop: 5,
-    paddingHorizontal: 30,
-    backgroundColor: 'transparent',
-    color: 'rgba(255, 255, 255, 0.75)',
-    fontSize: 13,
-    fontStyle: 'italic',
-    textAlign: 'center',
-  },
-  slider: {
-    marginTop: 15,
-    overflow: 'visible', // for custom animations
-  },
-  sliderContentContainer: {
-    paddingVertical: 10, // for custom animation
-  },
-  paginationContainer: {
-    paddingVertical: 8,
-  },
-  paginationDot: {
-    width: 8,
-    height: 8,
-    borderRadius: 4,
-    marginHorizontal: 8,
-  },
-});
